+++ src/main/java/org/jungppo/bambooforest/member/dto/MemberDto.java
@@ -1,4 +1,4 @@
-package org.jungppo.bambooforest.dto.member;
+package org.jungppo.bambooforest.member.dto;
 
 import java.time.LocalDateTime;
