+++ src/main/java/org/jungppo/bambooforest/repository/member/MemberRepository.java
@@ -1,6 +1,6 @@
 package org.jungppo.bambooforest.repository.member;
 
-import org.jungppo.bambooforest.entity.member.MemberEntity;
+import org.jungppo.bambooforest.member.domain.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
