+++ src/main/java/org/jungppo/bambooforest/repository/member/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
 package org.jungppo.bambooforest.repository.member;
 
-import org.jungppo.bambooforest.entity.member.RefreshTokenEntity;
+import org.jungppo.bambooforest.member.domain.RefreshTokenEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
