+++ src/main/java/org/jungppo/bambooforest/security/oauth2/CustomOAuth2LoginSuccessHandler.java
@@ -7,8 +7,8 @@ import java.io.IOException;
 
 import org.jungppo.bambooforest.entity.type.OAuth2Type;
 import org.jungppo.bambooforest.entity.type.RoleType;
+import org.jungppo.bambooforest.member.service.RefreshTokenService;
 import org.jungppo.bambooforest.security.jwt.JwtMemberClaim;
-import org.jungppo.bambooforest.service.member.RefreshTokenService;
 import org.jungppo.bambooforest.util.CookieUtils;
 import org.jungppo.bambooforest.util.JwtUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
