+++ src/main/java/org/jungppo/bambooforest/service/payment/PaymentService.java
@@ -10,9 +10,9 @@ import org.jungppo.bambooforest.dto.payment.PaymentSetupResponse;
 import org.jungppo.bambooforest.dto.paymentgateway.PaymentResponse;
 import org.jungppo.bambooforest.dto.paymentgateway.toss.TossPaymentRequest;
 import org.jungppo.bambooforest.entity.battery.BatteryItem;
-import org.jungppo.bambooforest.entity.member.MemberEntity;
 import org.jungppo.bambooforest.entity.payment.PaymentEntity;
 import org.jungppo.bambooforest.entity.type.PaymentStatusType;
+import org.jungppo.bambooforest.member.domain.MemberEntity;
 import org.jungppo.bambooforest.repository.member.MemberRepository;
 import org.jungppo.bambooforest.repository.payment.PaymentRepository;
 import org.jungppo.bambooforest.response.exception.battery.BatteryNotFoundException;
